requirements.txt
scikit-learn==0.24.2
matplotlib==3.3.3
apache-airflow==2.2.5
apache-airflow-providers-amazon==3.0.0